Wconversion fixes
A couple Wunsafe-loop-optimizations fixes are in here too.

+++ src/mesh/diva_io.C
@@ -232,7 +232,7 @@ void DivaIO::write_stream (std::ostream& out_file)
      */
     for(unsigned int e=0; e<the_mesh.n_elem(); e++)
       if (the_mesh.elem(e)->active())
-        for (unsigned int s=0; s<the_mesh.elem(e)->n_sides(); s++)
+        for (unsigned short s=0; s<the_mesh.elem(e)->n_sides(); s++)
           if (the_mesh.elem(e)->neighbor(s) == NULL)
             {
               const AutoPtr<Elem> side(the_mesh.elem(e)->build_side(s));
@@ -250,7 +250,7 @@ void DivaIO::write_stream (std::ostream& out_file)
      */
     for(unsigned int e=0; e<the_mesh.n_elem(); e++)
       if (the_mesh.elem(e)->active())
-        for (unsigned int s=0; s<the_mesh.elem(e)->n_sides(); s++)
+        for (unsigned short s=0; s<the_mesh.elem(e)->n_sides(); s++)
           if (the_mesh.elem(e)->neighbor(s) == NULL)
             {
               const AutoPtr<Elem> side(the_mesh.elem(e)->build_side(s));
